eliminate duplicate for assignFrameworkService()
