+++ v2/cf-nodejs-logging-support/src/lib/middleware/response-accessor.ts
@@ -1,16 +1,12 @@
-import ExpressService from "./framework-services/express";
-import Config from "../config/config";
-import RestifyService from "./framework-services/restify";
-import HttpService from "./framework-services/plainhttp";
-import ConnectService from "./framework-services/connect";
 import { IFrameworkService } from "./interfaces";
+import { assignFrameworkService } from "./utils";
 
 export default class ResponseAccessor {
     private static instance: ResponseAccessor;
     private frameworkService: IFrameworkService;
 
     private constructor() {
-        this.frameworkService = this.assignFrameworkService();
+        this.frameworkService = assignFrameworkService();
     }
 
     public static getInstance(): ResponseAccessor {
@@ -30,20 +26,5 @@ export default class ResponseAccessor {
 
     public finishLog(record: any) {
         return this.frameworkService.finishLog(record);
-    }
-
-    private assignFrameworkService(): IFrameworkService {
-        const framework = Config.getInstance().getFramework();
-        switch (framework) {
-            //insert your custom framework logger here
-            case "restify":
-                return new RestifyService();
-            case "nodejs-http":
-                return new HttpService();
-            case "connect":
-                return new ConnectService();
-            default:
-                return new ExpressService();
-        }
     }
 }
